requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
apispec==6.9.0
billiard==4.2.4
black==25.12.0
bleach==6.3.0
blinker==1.9.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Deprecated==1.3.1
deprecation==2.1.0
ecdsa==0.19.1
exceptiongroup==1.3.1
flake8==7.3.0
Flask==3.1.2
Flask-Admin==2.0.2
Flask-Limiter==4.1.1
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
flower==2.0.1
gevent==25.9.1
googleapis-common-protos==1.72.0
gotrue==2.12.4
greenlet==3.3.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.1
librt==0.7.3
limits==5.6.0
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==4.1.1
mccabe==0.7.0
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-flask==0.60b1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
postgrest==2.25.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
realtime==2.25.1
redis==7.1.0
requests==2.32.5
rsa==4.9.1
sentry-sdk==2.47.0
six==1.17.0
SQLAlchemy==2.0.45
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
tornado==6.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
webargs==8.7.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.3
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1
